+++ packages/ormed_sqlite/lib/src/sqlite_driver_registration_stub.dart
@@ -1,0 +1,3 @@
+import 'sqlite_driver_registration.dart';
+
+void registerSqliteProjectDriver(SqliteConnectionRegistrar registrar) {}